+++ src/test/scala/com/databricks/labs/mosaic/core/index/IndexSystemIDTest.scala
@@ -1,20 +1,21 @@
 package com.databricks.labs.mosaic.core.index
 
+import com.databricks.labs.mosaic.{BNG, H3}
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers._
 
 class IndexSystemIDTest extends AnyFunSuite {
 
     test("IndexSystemID creation from string") {
-        IndexSystemID("H3") shouldEqual H3
-        IndexSystemID("BNG") shouldEqual BNG
-        an[Error] should be thrownBy IndexSystemID("XYZ")
+        IndexSystemFactory.getIndexSystem("H3") shouldEqual H3
+        IndexSystemFactory.getIndexSystem("BNG") shouldEqual BNG
+        an[Error] should be thrownBy IndexSystemFactory.getIndexSystem("XYZ")
     }
 
     test("IndexSystemID getIndexSystem from ID") {
-        IndexSystemID.getIndexSystem(H3) shouldEqual H3IndexSystem
-        IndexSystemID.getIndexSystem(BNG) shouldEqual BNGIndexSystem
-        an[Error] should be thrownBy IndexSystemID.getIndexSystem(null)
+        IndexSystemFactory.getIndexSystem(H3.name) shouldEqual H3IndexSystem
+        IndexSystemFactory.getIndexSystem(BNG.name) shouldEqual BNGIndexSystem
+        an[Error] should be thrownBy IndexSystemFactory.getIndexSystem(null)
     }
 
 }