+++ src/test/scala/com/databricks/labs/mosaic/expressions/raster/RST_BandMetadataBehaviors.scala
@@ -23,7 +23,7 @@ trait RST_BandMetadataBehaviors extends QueryTest {
         val rasterDfWithBandMetadata = mocks
             .getNetCDFBinaryDf(spark)
             .withColumn("subdatasets", rst_subdatasets($"path"))
-            .withColumn("bleachingSubdataset", array_max(map_keys($"subdatasets")))
+            .withColumn("bleachingSubdataset", element_at($"subdatasets", "bleaching_alert_area"))
             .select(
               rst_bandmetadata($"bleachingSubdataset", lit(1))
                   .alias("metadata")
@@ -32,7 +32,7 @@ trait RST_BandMetadataBehaviors extends QueryTest {
         mocks
             .getNetCDFBinaryDf(spark)
             .withColumn("subdatasets", rst_subdatasets($"path"))
-            .withColumn("bleachingSubdataset", element_at(map_keys($"subdatasets"), 1))
+            .withColumn("bleachingSubdataset", element_at($"subdatasets", "bleaching_alert_area"))
             .createOrReplaceTempView("source")
 
         noException should be thrownBy spark.sql("""
@@ -42,7 +42,7 @@ trait RST_BandMetadataBehaviors extends QueryTest {
         noException should be thrownBy mocks
             .getNetCDFBinaryDf(spark)
             .withColumn("subdatasets", rst_subdatasets($"path"))
-            .withColumn("bleachingSubdataset", element_at(map_keys($"subdatasets"), 1))
+            .withColumn("bleachingSubdataset", element_at($"subdatasets", "bleaching_alert_area"))
             .select(
               rst_bandmetadata($"bleachingSubdataset", lit(1))
                   .alias("metadata")